초기 프로필 사진 수정시 오류 해결 (#76)
* fix: 초기 프로필 이미지를 수정의 경우 파일 삭제하지 않도록 수정

* test: 초기 프로필 이미지 수정 테스트코드 추가

* fix: 초기 프로필 이미지 경로가 null이 될 수 있음을 고려하여 수정

* test: 초기 프로필 이미지 경로가 null일때 테스트코드 추가

+++ src/main/java/com/example/solidconnection/siteuser/service/SiteUserService.java
@@ -88,7 +88,10 @@ public class SiteUserService {
         SiteUser siteUser = siteUserRepository.getByEmail(email);
         validateProfileImage(imageFile);
 
-        s3Service.deleteExProfile(email);
+        // 프로필 이미지를 처음 수정하는 경우에는 deleteExProfile 수행하지 않음
+        if(!isDefaultProfileImage(siteUser.getProfileImageUrl())){
+            s3Service.deleteExProfile(email);
+        }
         UploadedFileUrlResponse uploadedFileUrlResponse = s3Service.uploadFile(imageFile, ImgType.PROFILE);
         siteUser.setProfileImageUrl(uploadedFileUrlResponse.fileUrl());
         siteUserRepository.save(siteUser);
@@ -100,6 +103,11 @@ public class SiteUserService {
         if (imageFile == null || imageFile.isEmpty()) {
             throw new CustomException(PROFILE_IMAGE_NEEDED);
         }
+    }
+
+    private boolean isDefaultProfileImage(String profileImageUrl) {
+        String prefix = "https://solid-connection-uploaded.s3.ap-northeast-2.amazonaws.com/profile/";
+        return profileImageUrl == null || !profileImageUrl.startsWith(prefix);
     }
 
     /*
